move common training code into lib
