+++ src/binding/python/__init__.py
@@ -1,2 +1,2 @@
 from ._pyllmq import *
-from .wandb_log import setup_wandb, WandbLog
+from .training import TrainingConfig, CosineLRSchedule, training_logger_context
